dbeaver/pro#7458 backend now send event on canceling query job
